Introduce RenderDocumentHostUserData::render_frame_host accessor.
This CL refactors/deduplicates code by storing a RenderFrameHost*
pointer in a field of RenderDocumentHostUserData (and getting rid of an
equivalent field in some of the subclasses of RDHUD in favour of using
the `render_frame_host()` accessor instead - e.g. in FrameUsbServices,
MediaInterfaceProxy, and SerialService).

Bug: 1252044
Change-Id: I55d79846bed872e2bd84afe40cb7a0ee24dc1049
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3176183
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927334}