capture_mode_projector: Fix a flaky test
ProjectorCaptureModeIntegrationTests.RecordingOverlayWidgetBounds was
flaky. When it failed, a crash was observed due to a failure in writing
the remaining muxer frames when |RecordingEncoderMuxer| is being
destroyed on the encoder sequence.

The destruction of |RecordingEncoderMuxer| is asynchronous, and can
happen after the scoped temp dir
|MockProjectorClient::screencast_container_path_| gets destroyed,
causing a write failure.

|RecordingEncoderMuxer::on_failure_callback_| used to be declared
after |webm_muxer_|, and as a result, it used to get destroyed
first. So when an IO failure happens when flushing the muxer during
destruction happens and we attempt to invoke |on_failure_callback_|
(which was destroyed by then) we crash.

This CL fixes the declaration order.

BUG=1255090
TEST=Run the test and simulate an error, a crash should not be seen.

Change-Id: I5c012ec136312e19a5a32c0940d7b32a6715bb46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3198383
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927341}