evdev: Add methods to detect number pads to EventDeviceInfo
As part of work on collecting metrics for utilization of number
pads, we need to be able to identify input devices that are/have
numberpads.

This is preparation for further work, and does not itself change
any device behaviour.

Bug: b:184298258
Test: ./out/Debug/events_unittests --gtest_filter=EventDeviceInfoTest*
Change-Id: If5d921f83166da8b00473477e92dc3f819d58205
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3194378
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Auto-Submit: Kenneth Albanowski <kenalba@google.com>
Commit-Queue: Kenneth Albanowski <kenalba@google.com>
Cr-Commit-Position: refs/heads/main@{#927372}