Add Launcher nudge controller
For the discovery of the home button to encourage new users to press on
the icon and open the launcher app list, we implement the launcher
nudge on the home button as a feature.

This cl creates the launcher nudge controller class which is owned by
the ShelfController. The controller checks at the point a user is
logged in and decide whether the nudge should be shown or not and when
to show it. Note that there should be at most one
LauncherNudgeController at a time.

Bug: 1226093
Change-Id: Id1f52dd655a7a17db16a730e64737ed2cfdd5a41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3123427
Commit-Queue: Wen-Chien Wang <wcwang@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927413}