[DNR] Rename IndexHelper class to InstallIndexHelper
While working to defer the indexing of static rulesets, we renamed the
ReindexHelper class to IndexHelper. Confusingly however, there's
already another class called IndexHelper. With this change, we rename
the original IndexHelper to InstallIndexHelper and leave a TODO
referencing an issue for further work to combine the two classes.

Bug: 1231280
Change-Id: If9be05ffd9ae7ce73ddfa1c5adb0e732284f72d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3195913
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Commit-Queue: Dave Vandyke <kzar@kzar.co.uk>
Cr-Commit-Position: refs/heads/main@{#927420}