Fix drag from overflown bubble launcher to shelf
Fixes an issue where dragging from bubble launcher to parts of shelf
did not work when the bubble launcher's apps grid had overflow. Drag
events that were over parts of the shelf right under the bubble
launcher were not forwarded to shelf because the event location was
within the apps grid local bounds (even though that portion of local
bounds was clipped, and invisible to users), which was preventing
apps grid view from starting drag onto shelf.

After this cl, any events that drag and drop host can handle are sent to
the drag and drop host (even if the drag event is within the apps grid
local bounds) - this works as long as we ensure that shelf and launcher
bounds don't overlap, which is currently the case.
The change neccessitated an update in ApplicationDragAndDropHost
interface - changing ShouldStartDrag to ShouldHandleDrag (the method now
doesn not return false if a drag operation is already in progress,
though StartDrag has an extra check to ensure it doesn't start a new
drag operation in this case).

Change-Id: I2457ed7ca35a6297f70959ed24bfdc2f22524130
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3193110
Reviewed-by: Andrew Xu <andrewxu@chromium.org>
Commit-Queue: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927434}