[CrOS Bluetooth] Create DeviceNameManager.
Create DeviceNameManager to manage saving and retrieving Bluetooth
device nicknames. Internally, its implementation saves nicknames to
a prefs dictionary, with the key being the device's ID and the value
being the nickname. This nickname is visible for all users of the
device. This CL does not instantiate DeviceNameManager. This will
be done in a follow-up CL.

Bug: 1010321
Change-Id: Ia57312abdc38151149f95279ca46a4a72fed7041
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3191128
Auto-Submit: Gordon Seto <gordonseto@google.com>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927452}