[AppMenu] Mark skipEquality to true for AppMenuClickHandler
The problem is, when the view is reused, PropertyModel#compareValue uses
ObjectsCompat#equals to compare objects. It is good for other objects,
but for onItemClick(PropertyModel model), it only the method, not the
parameters, so even the 'model's are different, ObjectsCompat#equals
think they are same since both are #onItemClick.

This CL marks WritableObjectPropertyKey<AppMenuClickHandler>#skipEquality
to true, so when PropertyModel#compareValue to compare the model, it
always think they are different.

Bug: 1254786
Change-Id: Ib1d03f0d689d81ea1075b17457fc065d9138cd1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3199233
Commit-Queue: Gang Wu <gangwu@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927456}