[Autofill] Unset FrameData::parent_form of now-invisible frames.
This CL fixes a bug that was due to UpdateTreeOfRendererFrame()
erroneously assuming that frames removed from FormData::child_frames
have been or will be destroyed.

Frames may also be removed from FormData::child_frames because they
have become invisible.

In such a case, this CL unsets the removed frame's
FrameData::parent_form. That way, the former parent's
FormData::child_frames and the former child's FrameData::parent_form
are consistent.

As far as I can tell, this bug had no relevant consequences.

To keep the code simple, this CL does not push the fields that
originate from forms in the child's subtree to the child. The fields
are simply deleted, and only restored upon the next reparse of the
forms in the child's subtree.

Pushing the fields to the child's subtree is possible by tweaking
UpdateTreeOfRendererForm()'s tree traversal, but it does add some
further complexity.

Bug: 1187842, 1187857
Change-Id: I8fd7791e93609b1cf27cdc99eb917080c0f549c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3190490
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927548}