[ozone/wayland] Added a DCHECK on setting invalid surface scale.
The buffer scale in Wayland is integral and cannot be zero.  At the same
time, certain use cases such as forcing the device scale factor to
values between 0 and 1 result in that the surface gets a zero buffer
scale, which violates the Wayland protocol and terminates the connection
with the server.

This CL adds ceiling fractional scales when passing them to the surface,
so the buffer scale is always greater or equal to the DSF.

Bug: 1253604, 1241546
Change-Id: I9cd2ace0a6f6de2e2531edf87a02f2aef195e5ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3190730
Reviewed-by: Maksim Sisov <msisov@igalia.com>
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Cr-Commit-Position: refs/heads/main@{#927634}