Use synced dictionary in DevToolsSettings when sync is enabled
This CL updates `DevToolsSettings` to honor the boolean toggle that
allows users to explicitly opt-in to syncing DevTools settings.
Depending on the toggle, a different dictionary is used to store the
set of synced DevTools settings.

Please note that migrating and merging settings between the two
'synced' dictionaries when the toggle changes state is not handled
yet.

R=caseq@chromium.org

Bug: 1245541
Change-Id: Ie54ce84d98e8ebcc4dd52aa6221b306f25317a73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3189191
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927698}