Make LayerTreeHost::SourceFrameNumber() thread safe
There are a few places where SourceFrameNumber() is called on the impl
thread during commit. This has not been a problem, because the main
thread is blocked during commit, and SourceFrameNumber() is not
advanced until the end of commit.

However, when Non-Blocking Commit is implemented, the main thread and
impl thread will have different values for source frame number during
commit: the impl thread should use the frame number of the tree it is
synchronizing; while the main thread should use the frame number of the
*next* tree.

This patch makes code running on the impl thread use
LayerTreeHostImpl::sync_tree()->source_frame_number() everywhere; and
it ensures that LayerTreeImpl::set_source_frame_number() is called
early enough (in BeginCommit() rather than FinishCommitOnImplThread()).

Change-Id: If50bb1889ec9fed7b407eb3cb8632ac102c6ea69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3201033
Reviewed-by: vmpstr <vmpstr@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927756}