dPWAs: Move File Handling permission prompt to before window creation.
This change introduces a new flag to facilitate switching from
permissions to settings, called kDesktopPWAsFileHandlingSettingsGate.

When the flag is enabled, the File Handling permission flow now
matches the Protocol Handling permission flow in that a prompt is
shown before the app window is created rather than after. For now, as
the new prompt doesn't yet exist, it reuses the Protocol Handling
prompt. That prompt should be functionally equivalent, although it
lacks salient information such as which files and file types are at
play.

Some refactoring and renaming are also included in this patch since
some of the control flow is shared between FH and PH.

(Note this only handles Window and Linux --- Mac and CrOS are TODO.)

Bug: 1240696,1245289
Change-Id: Ib23081ff97155fb88ed820170f2062be9ce23cc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3192070
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Mike Jackson <mjackson@microsoft.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Auto-Submit: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927764}