[Page Zoom] Add page zoom behind experimental UI on Android
This CL continues the accessibility Page Zoom work behind a feature
flag.

Here we add the main functionality for page zoom and connect it with
the mock UI on Android. When the feature flags are enabled, the old
treatment of accessibility font scale factor will be disabled, and the
page zoom functionality will be leveraged instead. This uses the
existing zoom component and host zoom maps used on non-mobile
platforms. We add a simple singleton controller on the Java-side code
and connect it through JNI to the zoom component. The final resting
place for this UI is not yet decided, so the specifics of the
implementation can still change.

AX-Relnotes: N/A (Behind a flag)
Bug: 1232536, 645717
Change-Id: I1acc6dadcf1c70da44e1701f5e428e0fba92ce1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3166992
Reviewed-by: Bettina Dea <bdea@chromium.org>
Reviewed-by: Ehimare Okoyomon <eokoyomon@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Sean Topping <seantopping@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: W. James MacLean <wjmaclean@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#927772}