Diagnostics: add a simple keyboard diagram
This is the beginning of an interactive keyboard diagram to be used for
internal keyboards. In future it will show the appropriate keys and
labels for the device's physical keyboard and toggle their pressed state
when the user presses keys, but this CL lays out the basic grid-based
diagram.

Screenshot: http://shortn/_b5piuj30fb
(The "backspace" text overflowing is a known issue — not worth fixing
until the dialogue and diagram sizing is worked out later on.)

Bug: 1207678
Change-Id: Iea1a1594bacf803704be0e82b1520476b7abec1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3193505
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Commit-Queue: Jimmy Gong <jimmyxgong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927779}