Update MenuControllerCocoa to use ColorProvider for painting
This CL updates MenuControllerCocoa and related classes to take a
ColorProvider instance which is used to paint the NewTagImage
and IPHDotImage for Cocoa context menus.

This replaces the use of the global NativeTheme instance that
is in the process of being deprecated by the Views team. The
ColorProvider instance is taken from the Widget of the context
menu's source hierarchy where possible. Currently this is only
needed for render view context menus and Menus created in Views
via a MenuRunner. Note: Code should not hold references to a
give ColorProvider instance as they may be destroyed at
arbitrary times.

-[MenuControllerCocoa initWithModel:] has been updated to set
the instance variables but not initialize the menu
itself until -menu: is called. This allows creation of
the context menu to be deferred to when the ColorProvider
instance from the calling context is known.

This has been done since previously initWithModel would
be called in MenuRunnerImpl's constructor but the
ColorProvider is often not known during construction (but
is instead known when RunMenuAt() is called later on).

Bug: 1056756
Change-Id: I2a6caed5c0b4daf32c5e53f41ea7740719d88344
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3164141
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927791}