Add more accessibility unit tests on Android and fix focus edge case
This CL continues the expansion of the unit test suite for accessibility
on Android by adding more performAction tests.

With this CL we add tests for clearing and focusing a node, and giving
a node accessibility focus or clearing accessibility focus. We add
some clarifying comments to existing tests, and remove a dead method.
We also fix a rare edge case with the SET_PROGRESS action that would
prevent a user explicitly setting the progress of a slider to -1. We
fix an undiscovered caching issue when a new node is focused in Android
by clearing the AccessibilityNodeInfo cache for both the previously
focused node and the newly focused node, whereas to date we have
mistakenly only been clearing the newly focused node, which could be
producing rare and hard to reproduce bugs.

AX-Relnotes: N/A
Bug: N/A
Change-Id: Ib7ea938f87a588cab7406a70e023f55b18790078
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3199701
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927799}