Recommit presentation area into InkVisual when swap chain is recreated.
When the swap chain is recreated, it re-adds the Delegated Ink visual
to the Visual Tree, but, if the clipping rect was committed in a previous
step, this will not commit it again. The same happens to the ink color.

This happens on Edge when edge-pc-ui-integration is enabled.
In this scenario, there's a transparency on the window when it is
focused, and not when it isn't.
The swap chain is recreated when switching transparency modes.
(e.g. Alt+Tab).

When that happens, one sees the DelegatedInk painted with an offset from
the cursor in black (instead of expected color).

This CL fixes it by adding by setting a flag when the Swap Chain is
recreated, that is latter used when setting up the DelegatedInk and the
ink visual.

Bug: 1253632, 1253674, 1052145
Change-Id: I38c9035389d73f54c5552ebdbe77b1c10e52978f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3188731
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: João Victor Almeida de Aguiar <joalmei@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#927800}