Evaluate the end-line reshaping before adjusting to non-hangable-run-end
In r925962 we landed a patch to avoid the regression/crash reported in
https://crbug.com/1248710. The idea was to avoid creating the
line_start_result when the fist-safe position exceeded the value
of the non-hangable-run-end, instead of the actual break-opportunity
offset got from the new BreakIterator (we've changed its default
behavior to break after the whole space sequence).

However, this change caused different results in the evaluation of the
need of a line-end reshape, some lines bellow in this ShapeLine logic.
This change implied unneeded reshapes that lead to the performance
regression described in the bug 1254284.

The solution to avoid the perf regression was to move the logic to
evaluate the need of a line-end reshape before we adjust the breaking
opportunity offset to the non-hangable-run-end.

Finally, this CL is also fixing a typo in the crash-test added in
r925962, precisely to provide as regression test for the reported issue.


Bug: 1254284
Change-Id: I1cf747f3cfb4dd7c55950502ebdcc1be96d2f815
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3199791
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Javier Fernandez <jfernandez@igalia.com>
Cr-Commit-Position: refs/heads/main@{#927805}