Allow emergency-close of non-critical promo bubbles in a  screen region
Allows us to handle cases where the user triggers an IPH promo but then
interacts with a UI feature that would interact badly with the promo
bubble (without dismissing the bubble).

This causes serious rendering issues on Mac when, for example, the user
triggers the omnibox while an IPH bubble in the same area of the screen
is showing. (See attached bug.)

This CL adds a way to close non-critical promos overlapping a screen
region, and uses it to dismiss bubbles overlapping the Omnibox drop-down
when the user types into the location bar while a promo bubble is still
visible.

Bug: 1225046
Change-Id: I494afe7482d650564811fec771798d5dfa950ed5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3198199
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Emily Shack <emshack@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927814}