SpeechRecognitionPrivate: Error handling for start()
This is a follow-up to https://crrev.com/c/3145396. In
that change, we ironed out the details of how errors were
handled for the stop() function. This change follows up
and implements the same strategy for start(). Namely,
if start() is called and speech recognition is already
active, we should respond to the callback with an error.

Tests have been added/modified to verify the new behavior.

AX-Relnotes: N/A
Bug: 1246043

Change-Id: I9a6dadb171d78d7978f11b6dec344311a55a8e29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3171903
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927849}