desktop-share-menu: don't offer STTS if not available
In the future, STTS will always be "available" by having an onboarding
flow for its disabled cases, but that is not live yet; for now, don't
show it in the share submenu when it shouldn't be available. Without
this change, the STTS item in the menu DCHECKs on click.

Fixed: 1254523
Change-Id: Ibc30263ec7f09b7fc820578c0c712b3f548d4f72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3199653
Reviewed-by: Kristi Park <kristipark@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927861}