[SPC] CanMakePayment twice should return consistent results
Before the change, when two SPC PaymentRequest called CanMakePayment(),
the method returned different results. On Android, the first call will
hang. On Desktop, the first call will fail with false.

The issue behind is bug is that SPCAppFactory, a per-browser object
holds a unique pointer of InternalAuthenticator whose isUVPAA() is
called to calculate CanMakePayment. When only one CanMakePayment()
is called, there's no problem. InternalAuthenticator.isUVPAA(callback)
will be called can write its result in callback. When two
CanMakePayment() are called, there will be an issue of the second
InternalAuthenticator overriding the first InternalAuthenticator.
Since first InternalAuthenticator will have to be destructed. Since
the destruction on Android didn't resolve the pending callback, it
causes a hang on Android. Since the destruction on desktop resolves the
callback to false, it causes a result false on Desktop.

To fix this issue, this CL have SPCAppFactory no longer hold a unique
pointer of InternalAuthenticator, which was used to be referred to
in the isUVPAA callback before the CL. As a fix, this CL adds
InternalAuthenticator to Request and binds Request to isUVPAA's
callback, so that InternalAuthenticator will become one per request.

Also, the creation of Request was moved from OnIsUVPAAResponse()
to IsUVPAA() so that InternalAuthenticator can live in Request through
the async IsUVPAA().

After this CL, calling CanMakePayment() multiple times with SPC method
will be responded with consistent results.

Bug: 1229390
Change-Id: Idc9f6d32f472162922d9ca6f41453cd6374b63d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3194651
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927868}