[FilesSWA] Don't send FSP mount UMA metrics in background page for SWA
Files app SWA version doesn't have the concept of background pages and
the code that exists in the background page will eventually all live in
the foreground page. This means for multiple windows the mount metrics
will be duplicated, to avoid this stop tracking UMA when the SWA is
enabled. A follow up CL will be porting this logic to C++.

Bug: 1254243
Test: browser_tests --gtest_filter=FileManagerJsTest.MountMetrics*
Change-Id: Ic1ad2e8d85ccd54393bdede7dc9ac9f2c2bca898
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3195848
Reviewed-by: Bo Majewski <majewski@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927913}